# backend/core/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma de cinema screeK.",
        "title": "screeK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte screeK",
            "url": "http://www.screek.com/support",
            "email": "suporte@screek.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
...
        "/admin/management/cinemas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos os cinemas cadastrados (Apenas Admin/Manager)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Listar cinemas (Admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cinema.CinemaAdminResponseDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um novo cinema no sistema (Apenas Admin)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Cadastrar cinema",
                "parameters": [
                    {
                        "description": "Dados do cinema",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateCinemaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/management/cinemas/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados completos de um cinema, incluindo salas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Detalhes do cinema (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Cinema",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cinema.CinemaResponseDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Altera os dados de um cinema existente",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Atualizar cinema (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Cinema",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateCinemaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um cinema do sistema (Apenas se não houver salas vinculadas)",
                "tags": [
                    "Management"
                ],
                "summary": "Excluir cinema (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Cinema",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/management/cinemas/{id}/rooms": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adiciona uma sala a um cinema e gera o mapa de assentos automaticamente",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Criar sala e assentos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Cinema",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuração da sala",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/management/rooms/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Altera os dados da sala, mas não altera os assentos",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Atualizar sala (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sala",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados (CinemaID opcional aqui)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove uma sala, verificado se não há sessões futuras",
                "tags": [
                    "Management"
                ],
                "summary": "Excluir sala (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sala",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/management/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Consulta sessões de forma expandida para gestão",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Listar sessões administrativas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Cinema",
                        "name": "cinema_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cinema.SessionAdminResponseDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova sessão de filme em uma sala específica",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Agendar sessão",
                "parameters": [
                    {
                        "description": "Dados da sessão",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/management/sessions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Altera horário ou preço da sessão (Apenas se não houver ingressos)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Management"
                ],
                "summary": "Atualizar sessão (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sessão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cinema.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove uma sessão do sistema, desde que não existam ingressos vendidos",
                "tags": [
                    "Management"
                ],
                "summary": "Excluir sessão",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sessão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Room": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "cinema_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Seat"
                    }
                },
                "type": {
                    "$ref": "#/definitions/domain.RoomType"
                }
            }
        },
        "domain.RoomType": {
            "type": "string",
            "enum": [
                "STANDARD",
                "IMAX",
                "VIP"
            ],
            "x-enum-varnames": [
                "RoomTypeStandard",
                "RoomTypeIMAX",
                "RoomTypeVIP"
            ]
        },
        "domain.Seat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_occupied": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer"
                },
                "pos_x": {
                    "type": "integer"
                },
                "pos_y": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "row": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "httputil.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "cinema.CinemaAdminResponseDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "cinema.CinemaResponseDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Room"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "cinema.CreateCinemaRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "cinema.CreateRoomRequest": {
            "type": "object",
            "required": [
                "capacity",
                "cinema_id",
                "name",
                "type"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "cinema_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "STANDARD",
                        "IMAX",
                        "VIP"
                    ]
                }
            }
        },
        "cinema.CreateSessionRequest": {
            "type": "object",
            "required": [
                "movie_id",
                "price",
                "room_id",
                "session_type",
                "start_time"
            ],
            "properties": {
                "movie_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0
                },
                "room_id": {
                    "type": "integer"
                },
                "session_type": {
                    "type": "string",
                    "enum": [
                        "REGULAR",
                        "PREMIERE",
                        "RESCREENING",
                        "FESTIVAL",
                        "UNIVERSITY",
                        "SHOWCASE"
                    ]
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "cinema.SessionAdminResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "movie_title": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "room_name": {
                    "type": "string"
                },
                "session_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}